+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- [Provide an Oauth2 scope for the staff role #127](https://github.com/farmier/farm_rcd/issues/127)
+
 ## [1.2.0] 2026-01-21
 
 ### Added